backend/requirements/development.txt
-r ./_base.txt
coverage==6.4.1

# Code quality
django-stubs==1.8.0
djangorestframework-stubs==1.4.0
factory-boy==3.2.1
faker==8.12.1
flake8==3.9.2

# Debugging
ipdb==0.13.9
mypy==0.910
pylint==2.14.0
pylint-django==2.5.3

# Testing
pytest==6.2.5
pytest-django==4.4.0
pytest-sugar==0.9.4
